requirements.txt
llama-index==0.10.30
weaviate-client==4.6.5